refactor: add page controller
